requirements.txt
prettytable==3.8.0
python-dotenv==1.0.0
pytz==2022.6
Requests==2.31.0
